requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
bcrypt==3.2.0
bidict==0.21.2
broadcaster==0.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
constantly==15.1.0
cryptography==3.4.7
dataclasses==0.6
dateparser==1.0.0
dnspython==2.1.0
email-validator==1.1.2
Events==0.4
fastapi==0.61.1
fastapi-websocket-pubsub==0.1.18
fastapi-websocket-rpc==0.1.19
h11==0.9.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
iniconfig==1.1.1
mangum==0.12.2
motor==2.2.0
multidict==5.1.0
netifaces==0.10.6
packaging==21.0
passlib==1.7.4
pluggy==1.0.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.6.1
PyJWT==1.7.1
pymongo==3.11.3
pyOpenSSL==20.0.1
pyparsing==2.4.7
pytest==6.2.5
python-binance==1.0.10
python-binance-api==1.1.0
python-dateutil==2.8.1
python-decouple==3.4
python-engineio==4.2.0
python-socketio==5.3.0
pytz==2021.1
regex==2021.4.4
requests==2.25.1
service-identity==21.1.0
simplejson==3.17.2
six==1.15.0
starlette==0.13.6
tenacity==7.0.0
toml==0.10.2
Twisted==21.2.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
typing-extensions==3.10.0.0
tzlocal==2.1
ujson==4.0.2
urllib3==1.26.5
uvicorn==0.11.8
websocket-client==1.0.1
websockets==8.1
yarl==1.6.3
zope.interface==5.4.0
